Filter clusters heuristic

This removes the clusters from a fit object. There are 2 conditions that can be used to filter out clusters: 1) the cluster size (mixing proportion), which one can require to be above a certain cutoff, and 2) the Binomial peak, whichi one can ask to be above a certain value in at least a certain number of dimensions. Output clusters will be renamed by size (C1 will be larger etc.), and the latent variables and hard clustering assignments will be updated accordingly.

choose_clusters(x, binomial_cutoff = 0.05, dimensions_cutoff = 1,
  pi_cutoff = 0.02)

Arguments

x
An object of class 'vb_bmm'.

binomial_cutoff
Minimum Binomial success probability.

dimensions_cutoff
Minimum number of dimensions where we want to detect a Binomial component.

pi_cutoff
Minimum size of the mixture component.

Value

An object of class 'vb_bmm'.

Examples

data(fit_mvbmm_example)
choose_clusters(fit_mvbmm_example)

#> 
#> =-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-
#>  Selecting Binomial clusters (F1,2-heuristic).
#> =-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-
#> 
#> F1.       Cluster size 0.02 
#> F2. Num. of dimensions 1 
#> F2. Min. Binomial peak 0.05 
#> # A tibble: 10 x 7
#>    cluster            pi Above_cut F1    F2    accepted new.labels
#>    <chr>           <dbl>     <dbl> <lgl> <lgl> <lgl>    <chr>     
#>  1 C5      0.202                 2 TRUE  TRUE  TRUE     C1        
#>  2 C8      0.447                 2 TRUE  TRUE  TRUE     C2        
#>  3 C1      0.00000000433         2 FALSE TRUE  FALSE    C3        
#>  4 C2      0.00000000433         2 FALSE TRUE  FALSE    C4        
#>  5 C3      0.00000000433         2 FALSE TRUE  FALSE    C5        
#>  6 C4      0.00000000433         2 FALSE TRUE  FALSE    C6        
#>  7 C6      0.0693                1 TRUE  FALSE FALSE    C7        
#>  8 C7      0.00000000433         2 FALSE TRUE  FALSE    C8        
#>  9 C9      0.00000000433         2 FALSE TRUE  FALSE    C9        
#> 10 C10     0.281                 1 TRUE  FALSE FALSE    C10       

#>  [ mvbmm - Variational fit for Binomial mixtures ] 
#> 
#>     Points N = 231 
#> Dimensions W = 2 
#>        Fit CONVERGED (20 steps; eps. 1e-10) with Variational 
#> 
#> Binomial parameters (2 digits rounded) 
#>  # A tibble: 2 x 2
#>      C2    C1
#>   <dbl> <dbl>
#> 1  0.5   0.25
#> 2  0.49  0.25
#> 
#> Proportions (ordered)  
#> # A tibble: 2 x 2
#>   entry value
#>   <chr> <dbl>
#> 1 C2    0.688
#> 2 C1    0.312