Return clone trees from the fit.

This function uses the output fit of VIBER to create a call to ctree (https://caravagn.github.io/ctree/), a package to create clone trees for cancer evolution models.

Creation of a clone tree requires annotations that are not usually necessary for just a plain VIBER analyses. These annotations report the status of driver and gene for each one of the input datapoints; the annotations should have been passed when calling the variational_fit function, and stored inside the data field of the VIBER object.

The clonal cluster is estimated from the cluster with the highest parameter values in most of the input dimensions (Binomial peaks).

The output is the result of calling the constructor ctree::cetrees on the input clustering results x.

get_clone_trees(x, ...)

Arguments

x
A VIBER fit.

...
Extra parameters passed to the constructor ctree::cetrees, which affect the sampling of the trees.

Value

The output of the constructor ctree::cetrees.

Examples

data(mvbmm_example)

# We create annotation data assigning dummy names
# and picking 10 events to be drivers (randomly chosen)
data_annotations = data.frame(
 gene = paste0("G", 1:nrow(mvbmm_example$trials)),
 driver = FALSE
 )
 
data_annotations$driver[sample(1:nrow(data_annotations), 10)] = TRUE

# Compared to the main variational_fit, we use the same call but add data
f = variational_fit(mvbmm_example$successes, mvbmm_example$trials, data = data_annotations)

#>  [ VIBER - variational fit ] 
#> 

#> ℹ Input n = 231, with k < 10. Dirichlet concentration α = 1e-06.

#> ℹ Beta (a_0, b_0) = (1, 1); q_i = prior. Optimise: ε = 1e-10 or 5000 steps, r = 10 starts.

#> 

#> ✔ VIBER fit completed in 0.08 mins (status: converged)

#> 

#> ── [ VIBER ] My VIBER model n = 231 (w = 2 dimensions). Fit with k = 10 clusters

#> ● Clusters: π = 45% [C6], 28% [C9], 20% [C5], and 7% [C8], with π > 0.

#> ● Binomials: θ = <0.5, 0.49> [C6], <0, 0.2> [C9], <0.25, 0.25> [C5], and
#>   <0.22, 0> [C8].

#> ℹ Score(s): ELBO = -47073.328. Fit converged in 26 steps, ε = 1e-10.

print(f)

#> ── [ VIBER ] My VIBER model n = 231 (w = 2 dimensions). Fit with k = 10 clusters

#> ● Clusters: π = 45% [C6], 28% [C9], 20% [C5], and 7% [C8], with π > 0.

#> ● Binomials: θ = <0.5, 0.49> [C6], <0, 0.2> [C9], <0.25, 0.25> [C5], and
#>   <0.22, 0> [C8].

#> ℹ Score(s): ELBO = -47073.328. Fit converged in 26 steps, ε = 1e-10.

trees = get_clone_trees(f)

#> Error in (x$N * x$pi_k) %>% round: could not find function "%>%"

ctree:::print.ctree(trees[[1]])

#> Error in loadNamespace(name): there is no package called ‘ctree’

ctree::plot.ctree(trees[[1]])

#> Error in loadNamespace(name): there is no package called ‘ctree’