Variational fit for Binomial mixtures

Variational fit for a semi-parametric Dirichelt mixture of Binomial distributions. The fit convergency can be monitored through the ELBO, can be run either sequentially (single core) or in parallel. You need to provide an upper bound on the number of clusters that you want to obtain, through parameters K. You can explicit the Dirichlet prior for the concentration of the mixture (alpha_0), as well as the hyperparmeters of the Beta priors for each mixture component.

variational_fit(x, y, K = 10, alpha_0 = 1e-06, a_0 = 1, b_0 = 1,
  max_iter = 5000, epsilon_conv = 1e-10, samples = 10,
  q_init = "prior", trace = FALSE)

Arguments

x
A matrix where each column is a dimension of the multivariate Binomial, and each row is an input point. Values of this matrix represent the number of successes of independent Bernoulli trials. This matrix and y should have the same dimension.

y
A matrix where each column is a dimension of the multivariate Binomial, and each row is an input point. Values of this matrix represent the number of attempts of independent Bernoulli trials. This matrix and x should have the same dimension.

K
The maximum number of clusters returned, it should be lower than the number of rows of x and y. Default is K = 10; lower values speed up convergence.

alpha_0
The concentration parameter of the Dirichlet mixture. The default is a stringent fit with alpha = 1e-6.

a_0
Prior Beta hyperparameter. If this values is a scalar than all the mixture components have the same prior. The default is scalar a_0 = 1.

b_0
Prior Beta hyperparameter. If this values is a scalar than all the mixture components have the same prior. The default is scalar b_0 = 1.

max_iter
Maximum number of fit iterations. The fit is interrupted when this number of iterations is performed. Default max_iter = 5000

epsilon_conv
Epsilon to measure convergence (ELBO absolute difference).

samples
Number of fits computed by the algorithm. Only the best fit is returned. This value must be greater or equal than 1.

q_init
Initialization of the q-distribution to compute the approximation of the posterior distributions. This can be set in three different waysL equal to the prior (q_init = 'prior'), via kmeans clustering (q_init = 'kmeans') and capturing points which are private to each dimension (q_init = 'private'). The default is equal to the prior.

trace
If true the trace computed during the fit is returned (this allows to check fits a posterirori, make animations etc.). Default is FALSE; this feature can slow down quite substantially the fit.

Value

An object of class vb_bmm which contains S3 methods to extract the fit, plots the results, compute summary statistics etc.

Examples

data(mvbmm_example)
f = variational_fit(mvbmm_example$successes, mvbmm_example$trials)

#>  [ VIBER - variational fit ] 
#> 
#> 	INPUT
#>   Points N = 231 
#> Clusters K = 10 (max)
#> 
#> Dirichlet alpha = 1e-06 (conc.)
#>      Beta a0 = 1; b0 =1 (shape)
#> 
#>      Beta (posterior) prior 
#>  Optimize epsilon = 1e-10; steps =5000; r = 10 
#> 
#> 
#> 
#> BEST FIT
#> 
#>  [ VIBER - Variational inference for Binomial mixtures ] 
#> 
#>     Points N = 231 
#> Dimensions W = 2 
#>        Fit CONVERGED (20 steps; eps. 1e-10) with Variational 
#> 
#> Binomial parameters (2 digits rounded) 
#>  # A tibble: 2 x 10
#>      C6    C8    C5    C4    C1    C2    C3    C7    C9   C10
#>   <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl>
#> 1  0.5    0    0.25  0.22   0.5   0.5   0.5   0.5   0.5   0.5
#> 2  0.49   0.2  0.25  0      0.5   0.5   0.5   0.5   0.5   0.5
#> 
#> Proportions (ordered)  
#> # A tibble: 10 x 2
#>    entry         value
#>    <chr>         <dbl>
#>  1 C6    0.447        
#>  2 C8    0.281        
#>  3 C5    0.202        
#>  4 C4    0.0693       
#>  5 C1    0.00000000433
#>  6 C2    0.00000000433
#>  7 C3    0.00000000433
#>  8 C7    0.00000000433
#>  9 C9    0.00000000433
#> 10 C10   0.00000000433
#> 
#> 
#> COMPLETED: 0.1 mins, with status CONVERGED 
#> 

print(f)

#>  [ VIBER - Variational inference for Binomial mixtures ] 
#> 
#>     Points N = 231 
#> Dimensions W = 2 
#>        Fit CONVERGED (20 steps; eps. 1e-10) with Variational 
#> 
#> Binomial parameters (2 digits rounded) 
#>  # A tibble: 2 x 10
#>      C6    C8    C5    C4    C1    C2    C3    C7    C9   C10
#>   <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl>
#> 1  0.5    0    0.25  0.22   0.5   0.5   0.5   0.5   0.5   0.5
#> 2  0.49   0.2  0.25  0      0.5   0.5   0.5   0.5   0.5   0.5
#> 
#> Proportions (ordered)  
#> # A tibble: 10 x 2
#>    entry         value
#>    <chr>         <dbl>
#>  1 C6    0.447        
#>  2 C8    0.281        
#>  3 C5    0.202        
#>  4 C4    0.0693       
#>  5 C1    0.00000000433
#>  6 C2    0.00000000433
#>  7 C3    0.00000000433
#>  8 C7    0.00000000433
#>  9 C9    0.00000000433
#> 10 C10   0.00000000433